requirements.txt
flask==3.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
zeep==4.2.1
lxml==5.1.0
pyyaml==6.0.1
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
